♻️ refactor: instance 재설정
