+++ app/_apis/shop/index.ts
@@ -1,4 +1,3 @@
-import axiosInstance from "../instances";
 import instance from "@/app/_lib/axios";
 import { AddressType } from "@/app/_constants/address";
 import { CategoryType } from "@/app/_constants/category";
@@ -52,7 +51,7 @@ export const postShopCreate = async ({
 }): Promise<boolean> => {
   try {
     const processedImageUrl = await getImageUrl(imageUrl);
-    const response = await axiosInstance.post<PostShop>(
+    const response = await instance.post<PostShop>(
       "/shops",
       {
         name,
