+++ app/_lib/axios.ts
@@ -1,11 +1,34 @@
 import axios from "axios";
+import { getCookie } from "../_util/cookie";
+
+declare module "axios" {
+  export interface AxiosRequestConfig {
+    authorization?: boolean;
+  }
+}
 
 const instance = axios.create({
-  baseURL: process.env.NEXT_PUBLIC_BASE_URL,
-  headers: {
-    "Content-Type": "application/json",
-    Authorization: `Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiIzMDlhYWY2Mi0wNjhlLTRkZWItYTZjMy1hMzFhYmFjZmRjNjciLCJpYXQiOjE3MTk4MDQ5NDh9.t9C0zBJupLJxfLPxXnYHlayXzXshEMIidHmUY7tjyS0`,
-  },
+  baseURL: process.env.NEXT_PUBLIC_GHEUP_PAY_BASE_URL,
+  timeout: 20000,
+  authorization: true,
 });
+
+instance.interceptors.request.use(
+  async (config) => {
+    const hasConfig =
+      !config.authorization || !config.headers || config.headers.Authorization;
+    if (hasConfig) {
+      return config;
+    }
+    const accessToken = await getCookie("accessToken");
+
+    if (!accessToken) return config;
+    config.headers.Authorization = `Bearer ${accessToken}`;
+    return config;
+  },
+  (error) => {
+    return Promise.reject(error);
+  },
+);
 
 export default instance;
